Added instruction and script files
Added the main assignment document that contains all the instructions for completing the assignment. Also added the script files that create database tables, add constraints and load data in the tables. There is a WKIS_ERD.pdf that contains the physical model for the database used in this assignment. All these files were provided by the instructor.